fix: :recycle: move showSelectedList function to its own module (is used in 2 places now)
hopefully this doesn't break everything
